refactor(upload conflicts): Move conflict types to a types file PE-639
